some changes requested
