+++ .gitignore
@@ -1,1 +1,1 @@
-src/__pycache__
+src/is_face_detector/__pycache__
